course2/requirements.txt

tqdm 
PyYaml
opencv-python 
numpy 
Pillow 

torchmetrics
torch==1.10.2+cu113 
torchvision==0.11.3+cu113
-f https://download.pytorch.org/whl/cu113/torch_stable.html